Rename Jetty and Tomcat common modules
