+++ .fossa.yml
@@ -606,7 +606,7 @@ targets:
       target: ':instrumentation:jetty:jetty-8.0:javaagent'
     - type: gradle
       path: ./
-      target: ':instrumentation:jetty:jetty-common:javaagent'
+      target: ':instrumentation:jetty:jetty-common-8.0:javaagent'
     - type: gradle
       path: ./
       target: ':instrumentation:jetty-httpclient:jetty-httpclient-12.0:javaagent'
@@ -1029,7 +1029,7 @@ targets:
       target: ':instrumentation:tomcat:tomcat-7.0:javaagent'
     - type: gradle
       path: ./
-      target: ':instrumentation:tomcat:tomcat-common:javaagent'
+      target: ':instrumentation:tomcat:tomcat-common-7.0:javaagent'
     - type: gradle
       path: ./
       target: ':instrumentation:tomcat:tomcat-jdbc:javaagent'
