+++ instrumentation/jetty/jetty-11.0/javaagent/build.gradle.kts
@@ -14,7 +14,7 @@ muzzle {
 dependencies {
   library("org.eclipse.jetty:jetty-server:11.0.0")
 
-  implementation(project(":instrumentation:jetty:jetty-common:javaagent"))
+  implementation(project(":instrumentation:jetty:jetty-common-8.0:javaagent"))
   implementation(project(":instrumentation:servlet:servlet-5.0:javaagent"))
   bootstrap(project(":instrumentation:servlet:servlet-common:bootstrap"))
