+++ instrumentation/jetty/jetty-8.0/javaagent/build.gradle.kts
@@ -17,7 +17,7 @@ dependencies {
 
   library("org.eclipse.jetty:jetty-server:8.0.0.v20110901")
 
-  implementation(project(":instrumentation:jetty:jetty-common:javaagent"))
+  implementation(project(":instrumentation:jetty:jetty-common-8.0:javaagent"))
   implementation(project(":instrumentation:servlet:servlet-3.0:javaagent"))
   bootstrap(project(":instrumentation:servlet:servlet-common:bootstrap"))
