+++ instrumentation/tomcat/README.md
@@ -2,7 +2,7 @@
 
 Tomcat support is divided into the following sub-modules:
 
-- `tomcat-common:javaagent` contains common type instrumentation, advice helper classes and abstract
+- `tomcat-common-7.0:javaagent` contains common type instrumentation, advice helper classes and abstract
   tracer used by the `javaagent` modules of all supported Tomcat versions
 - `tomcat-7.0:javaagent` applies Tomcat request handler instrumentation for versions `[7, 10)`
 - `tomcat-10.0:javaagent` applies Tomcat request handler instrumentation for versions `[10,)`
