+++ instrumentation/tomcat/tomcat-10.0/javaagent/build.gradle.kts
@@ -12,7 +12,7 @@ muzzle {
 }
 
 dependencies {
-  implementation(project(":instrumentation:tomcat:tomcat-common:javaagent"))
+  implementation(project(":instrumentation:tomcat:tomcat-common-7.0:javaagent"))
   implementation(project(":instrumentation:servlet:servlet-5.0:javaagent"))
   bootstrap(project(":instrumentation:servlet:servlet-common:bootstrap"))
