+++ instrumentation/tomcat/tomcat-10.0/javaagent/src/main/java/io/opentelemetry/javaagent/instrumentation/tomcat/v10_0/Tomcat10Singletons.java
@@ -8,8 +8,8 @@ package io.opentelemetry.javaagent.instrumentation.tomcat.v10_0;
 import io.opentelemetry.instrumentation.api.instrumenter.Instrumenter;
 import io.opentelemetry.instrumentation.servlet.v5_0.internal.Servlet5Accessor;
 import io.opentelemetry.javaagent.instrumentation.servlet.v5_0.Servlet5Singletons;
-import io.opentelemetry.javaagent.instrumentation.tomcat.common.TomcatHelper;
-import io.opentelemetry.javaagent.instrumentation.tomcat.common.TomcatInstrumenterFactory;
+import io.opentelemetry.javaagent.instrumentation.tomcat.common.v7_0.TomcatHelper;
+import io.opentelemetry.javaagent.instrumentation.tomcat.common.v7_0.TomcatInstrumenterFactory;
 import jakarta.servlet.http.HttpServletRequest;
 import jakarta.servlet.http.HttpServletResponse;
 import org.apache.coyote.Request;
