+++ instrumentation/tomcat/tomcat-7.0/javaagent/src/main/java/io/opentelemetry/javaagent/instrumentation/tomcat/v7_0/Tomcat7Singletons.java
@@ -8,8 +8,8 @@ package io.opentelemetry.javaagent.instrumentation.tomcat.v7_0;
 import io.opentelemetry.instrumentation.api.instrumenter.Instrumenter;
 import io.opentelemetry.instrumentation.servlet.v3_0.internal.Servlet3Accessor;
 import io.opentelemetry.javaagent.instrumentation.servlet.v3_0.Servlet3Singletons;
-import io.opentelemetry.javaagent.instrumentation.tomcat.common.TomcatHelper;
-import io.opentelemetry.javaagent.instrumentation.tomcat.common.TomcatInstrumenterFactory;
+import io.opentelemetry.javaagent.instrumentation.tomcat.common.v7_0.TomcatHelper;
+import io.opentelemetry.javaagent.instrumentation.tomcat.common.v7_0.TomcatInstrumenterFactory;
 import javax.servlet.http.HttpServletRequest;
 import javax.servlet.http.HttpServletResponse;
 import org.apache.coyote.Request;
