+++ instrumentation/tomcat/tomcat-common-7.0/javaagent/src/main/java/io/opentelemetry/javaagent/instrumentation/tomcat/common/v7_0/TomcatRequestGetter.java
@@ -3,7 +3,7 @@
  * SPDX-License-Identifier: Apache-2.0
  */
 
-package io.opentelemetry.javaagent.instrumentation.tomcat.common;
+package io.opentelemetry.javaagent.instrumentation.tomcat.common.v7_0;
 
 import static java.util.Collections.emptyIterator;
