+++ settings.gradle.kts
@@ -400,7 +400,7 @@ include(":instrumentation:jedis:jedis-common-1.4:javaagent")
 include(":instrumentation:jetty:jetty-8.0:javaagent")
 include(":instrumentation:jetty:jetty-11.0:javaagent")
 include(":instrumentation:jetty:jetty-12.0:javaagent")
-include(":instrumentation:jetty:jetty-common:javaagent")
+include(":instrumentation:jetty:jetty-common-8.0:javaagent")
 include(":instrumentation:jetty-httpclient:jetty-httpclient-9.2:javaagent")
 include(":instrumentation:jetty-httpclient:jetty-httpclient-9.2:library")
 include(":instrumentation:jetty-httpclient:jetty-httpclient-9.2:testing")
@@ -698,7 +698,7 @@ include(":instrumentation:thrift-0.13:library")
 include(":instrumentation:thrift-0.13:testing")
 include(":instrumentation:tomcat:tomcat-7.0:javaagent")
 include(":instrumentation:tomcat:tomcat-10.0:javaagent")
-include(":instrumentation:tomcat:tomcat-common:javaagent")
+include(":instrumentation:tomcat:tomcat-common-7.0:javaagent")
 include(":instrumentation:tomcat:tomcat-jdbc:javaagent")
 include(":instrumentation:twilio-6.6:javaagent")
 include(":instrumentation:undertow-1.4:bootstrap")
